move payloadinterface and variant to segment
